% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDebuggingPanelOutput.R
\name{makeDebuggingPanelOutput}
\alias{makeDebuggingPanelOutput}
\title{makeDebuggingPanelOutput}
\usage{
makeDebuggingPanelOutput(session)
}
\description{
Place shinyDebuggingPanel::makeDebuggingPanelOutput(session) inside your call to shinyServer(server.R),
Place shinyDebuggingPanel::withdebuggingPanel() in your ui.R
}

