% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDebuggingPanelOutput.R, R/makeDebuggingPanelOutput.master.R
\name{makeDebuggingPanelOutput}
\alias{makeDebuggingPanelOutput}
\title{makeDebuggingPanelOutput}
\usage{
makeDebuggingPanelOutput(session)

makeDebuggingPanelOutput(session)
}
\description{
Place shinyDebuggingPanel::makeDebuggingPanelOutput(session) inside your call to shinyServer(server.R),
Place shinyDebuggingPanel::withDebuggingPanel() in your ui.R

Place shinyDebuggingPanel::makeDebuggingPanelOutput(session) inside your call to shinyServer(server.R),
Place shinyDebuggingPanel::withDebuggingPanel() in your ui.R
}
\details{
Make sure your server function has a third argument:  function(input, output, session)

Make sure your server function has a third argument:  function(input, output, session)
}

