% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDebuggingPanelOutput.R
\name{makeDebuggingPanelOutput}
\alias{makeDebuggingPanelOutput}
\title{makeDebuggingPanelOutput}
\usage{
makeDebuggingPanelOutput(session)
}
\arguments{
\item{session}{The shiny session; default is  shiny::getDefaultReactiveDomain()s}
}
\description{
Place shinyDebuggingPanel::makeDebuggingPanelOutput(session) inside your call to shinyServer(server.R). Place shinyDebuggingPanel::withDebuggingPanel() in your ui.R.
}
\details{
Make sure your server function has the third argument:  function(input, output, session)
}

